migrate using-mdx example
